Fix Digest Mismatch error on Enterprise Linux installations when FIPS mode enabled
Added fpm argument to the electron-builder.json file to change the hashing algorithm from md5 to sha256. This fixes an issue introduced in RHEL8+ and/or clones where if FIPS mode is enabled rpm digests must be hashed with at minimum sha256 to be allowed to be installed without bypassing security measures put in place by the FIPS standards.